\name{packtest2-package}
\alias{packtest2-package}
\alias{packtest2}
\docType{package}
\title{
\packageTitle{packtest2}
}
\description{
\packageDescription{packtest2}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{packtest2}
\packageIndices{packtest2}
%~~ An overview of how to use the package, including the most important ~~
%~~ functions ~~
}
\author{
\packageAuthor{packtest2}

Maintainer: \packageMaintainer{packtest2}
}
\references{
%~~ Literature or other references for background information ~~
}
%~~ Optionally other standard keywords, one per line, from file KEYWORDS in the ~~
%~~ R documentation directory ~~
\keyword{ package }
\seealso{
%~~ Optional links to other man pages, e.g. ~~
%~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%~~ simple examples of the most important functions ~~
}
